feat: settings for disable/enable mentions

+++ core/bot.ts
@@ -1,5 +1,5 @@
 import { Context, NarrowedContext, Telegraf } from 'telegraf'
-import { UserRepository } from './repository.js'
+import { UserRepository } from './userRepository.js'
 import { MetricsService } from './metrics.js'
 import { message } from 'telegraf/filters'
 import {
@@ -14,6 +14,8 @@ import {
   matchMentionsToUsers,
   getMentionsFromEntities,
   isChatGroup,
+  createSettingsKeyboard,
+  Settings,
 } from './utils/utils.js'
 import { MentionRepository } from './mentionRepository.js'
 import {
@@ -28,10 +30,13 @@ import {
   NEED_TO_BUY_UNLIMITED,
   NEW_MENTION_EXAMPLE,
   NOT_EXISTED_MENTION_TEXT,
+  ONLY_ADMIN_ACTION_TEXT,
   PRIVACY_COMMAND_TEXT,
+  SETTINGS_TEXT,
 } from './constants/texts.js'
 import { LIMITS_MENTION_FOR_ADDING_PAY } from './constants/limits.js'
 import { PaymentsRepository } from './paymentsRepository.js'
+import { SettingsAction, SettingsRepository } from './settingsRepository.js'
 
 type UniversalMessageOrActionUpdateCtx = NarrowedContext<
   Context,
@@ -70,6 +75,7 @@ export class Bot {
     private readonly metricsService: MetricsService,
     private readonly mentionRepository: MentionRepository,
     private readonly paymentsRepository: PaymentsRepository,
+    private readonly settingsRepository: SettingsRepository,
     botName: string | undefined,
     token: string | undefined
   ) {
@@ -107,6 +113,10 @@ export class Bot {
         description: '/delete_mention some_group',
       },
       {
+        command: 'settings',
+        description: 'Bot settings',
+      },
+      {
         command: 'help',
         description: 'Help information',
       },
@@ -114,18 +124,15 @@ export class Bot {
         command: 'donate',
         description: 'Support the project to pay for servers and new features',
       },
-      // Disable for smallest commands line
-      // {
-      //   command: 'privacy',
-      //   description: 'How the Bot takes care of your personal data',
-      // },
-      // TODO rename mention?
     ])
 
     this.registerDonateCommand()
     this.registerPrivacyCommand()
     this.registerHelpCommand()
 
+    this.registerSettingsChangeAction()
+
+    this.registerSettingsCommand()
     this.registerMentionCommand()
     this.registerGetAllMentionCommand()
     this.registerAddToMentionCommand()
@@ -275,6 +282,70 @@ ${INTRODUCE_CUSTOM_MENTIONS_TEXT}${
     this.metricsService.updateLatestPaymentsCall(`${chatId}`)
 
     return DONATE_COMMAND_TEXT
+  }
+
+  private registerSettingsChangeAction(): void {
+    this.bot.action(/^[/settings]+(_.+)?$/, async (ctx) => {
+      if (!ctx.chat?.id) return
+
+      const isAllowed = await this.getIsAllowed(
+        ctx.chat.id,
+        ctx.update.callback_query.from.id
+      )
+      if (!isAllowed) {
+        this.metricsService.restrictedAction.inc({
+          chatId: ctx.chat.id,
+          action: 'settingsChanged',
+        })
+
+        await ctx.reply(`<strong>🛑 Only admins can edit settings</strong>`, {
+          parse_mode: 'HTML',
+        })
+        return
+      }
+
+      const data = (ctx.update.callback_query as any).data as string
+
+      const [, jsonData] = data.split('_')
+
+      const { s, ...rest } = JSON.parse(jsonData) as Settings
+
+      const newValue = !rest[s]
+
+      const changedSettings = {
+        ...rest,
+        [s]: newValue,
+      }
+
+      await this.settingsRepository.updateSettings(ctx.chat.id, s, newValue)
+
+      const keyboard = createSettingsKeyboard(changedSettings)
+
+      await ctx.editMessageReplyMarkup({
+        inline_keyboard: keyboard,
+      })
+    })
+  }
+
+  private registerSettingsCommand(): void {
+    this.bot.command('settings', async (ctx) => {
+      const settings = await this.settingsRepository.getSettingsCompressed(
+        ctx.chat.id
+      )
+      const keyboard = createSettingsKeyboard(settings)
+
+      this.metricsService.commandsCounter.inc({
+        chatId: ctx.chat.id.toString(),
+        command: 'settings',
+      })
+
+      ctx.reply(SETTINGS_TEXT, {
+        parse_mode: 'HTML',
+        reply_markup: {
+          inline_keyboard: keyboard,
+        },
+      })
+    })
   }
 
   private registerMentionCommand(): void {
@@ -432,6 +503,12 @@ ${INTRODUCE_CUSTOM_MENTIONS_TEXT}${
         return
       }
 
+      const isAllowed = await this.onlyAdminActionGuard(
+        ctx,
+        'crudCustomMention'
+      )
+      if (!isAllowed) return
+
       const mentions = getMentionsFromEntities(
         ctx.message.text,
         ctx.message.entities
@@ -556,6 +633,12 @@ Need more? Try removing useless mentions using the /mentions and /delete_mention
         return
       }
 
+      const isAllowed = await this.onlyAdminActionGuard(
+        ctx,
+        'crudCustomMention'
+      )
+      if (!isAllowed) return
+
       const mentions = getMentionsFromEntities(
         ctx.message.text,
         ctx.message.entities
@@ -659,6 +742,12 @@ Contact us via support chat from /help`,
 
         return
       }
+
+      const isAllowed = await this.onlyAdminActionGuard(
+        ctx,
+        'crudCustomMention'
+      )
+      if (!isAllowed) return
 
       const wasDeleted = await this.mentionRepository.deleteMention(
         ctx.message.chat.id,
@@ -823,15 +912,15 @@ Contact us via support chat from /help`,
       )
       if (!isCallAll) return
 
+      const isAllowed = await this.onlyAdminActionGuard(ctx, 'useAllMention')
+      if (!isAllowed) return
+
       const chatUsernames = await this.userRepository.getUsernamesByChatId(
         chatId
       )
       const usernames = Object.values(chatUsernames).filter(
         (username) => username !== from.username
       )
-      // const usernames = new Array<string>(255)
-      //   .fill('item')
-      //   .map((_, index) => `test${index}`)
 
       if (!usernames.length) {
         console.log('[ALL] Noone to mention', ctx.message.chat.id)
@@ -1014,12 +1103,6 @@ Someone should write something (read more /help).
               buttons,
             ]
 
-            // await new Promise((resolve) => {
-            //   setTimeout(() => {
-            //     resolve(null)
-            //   }, 10_000)
-            // })
-
             try {
               await ctx.reply(lastStr, {
                 reply_to_message_id: withReply ? messageId : undefined,
@@ -1093,6 +1176,9 @@ Someone should write something (read more /help).
     field: string
   ): Promise<void> {
     if (!ctx.chat?.id) return
+
+    const isAllowed = await this.onlyAdminActionGuard(ctx, 'useCustomMention')
+    if (!isAllowed) return
 
     const ids = await this.mentionRepository.getUsersIdsByMention(
       ctx.chat.id,
@@ -1176,6 +1262,33 @@ Someone should write something (read more /help).
     }).catch(this.handleSendMessageError)
   }
 
+  private async onlyAdminActionGuard(
+    ctx: UniversalMessageOrActionUpdateCtx,
+    action: SettingsAction
+  ): Promise<boolean> {
+    if (!ctx.chat?.id) return true
+
+    const settings = await this.settingsRepository.getSettings(ctx.chat?.id)
+
+    const isActionOnlyForAdmin = settings[action]
+
+    const isAllowed =
+      !isActionOnlyForAdmin ||
+      (await this.getIsAllowed(ctx.chat?.id, ctx.from?.id))
+
+    if (isAllowed) return true
+
+    this.metricsService.restrictedAction.inc({
+      chatId: ctx.chat.id,
+      action,
+    })
+
+    await ctx.reply(ONLY_ADMIN_ACTION_TEXT, {
+      parse_mode: 'HTML',
+    })
+    return false
+  }
+
   private async getKeyboardWithCustomMentions(
     chatId: Chat['id']
   ): Promise<InlineKeyboardMarkup | undefined> {
@@ -1199,5 +1312,20 @@ Someone should write something (read more /help).
     })
 
     return keyboard
+  }
+
+  private async getIsAllowed(
+    chatId: Chat['id'],
+    userId: User['id'] | undefined
+  ): Promise<boolean> {
+    if (!userId) return true
+
+    const member = await this.bot.telegram.getChatMember(chatId, userId)
+    const allowed = [
+      'administrator',
+      'creator', // disable while debug
+    ]
+
+    return allowed.includes(member.status)
   }
 }
